refactor(web): chat UI の微調整と app/chat/components への移設
- SpeechBubble.tsx: 三角（テール）SVG を撤廃
- TopBar.tsx / Thread.tsx / ToolFallback.tsx: 任意値 text-[Npx] を
  Tailwind スケール (text-base / text-xs) に置換
- Constellation.tsx + index.css: 雪景色背景 (bg-winter.png) でも目立つ
  よう色・サイズを調整 (dot 1.2-2.6 → 2.0-3.6px、線幅 0.8 → 1.2、
  z-index 0 → 1、トークンを teal-700 系の薄め透明度に更新)
- chat 専用パーツ (Mascot / ChatStandingMascot / AmbientBG /
  Constellation / SpeechBubble / TopBar) を web/src/components/
  companion/ から web/src/app/chat/components/ へ移設し、
  FeedbackModal.tsx と並列に整理。companion フォルダは廃止
- 関連 import パスを更新
