+++ web/src/app/chat/components/AmbientBG.tsx
