+++ web/src/app/chat/components/ChatStandingMascot.tsx
