+++ web/src/app/chat/components/Mascot.tsx
