+++ web/src/index.css
@@ -176,10 +176,11 @@
 
   /* ===============================================
      CONSTELLATION (背景の星座レイヤー)
+     bg-winter.png (雪景色の白系) の上で目立つよう濃い teal で描画
      =============================================== */
-  --constellation-line: rgba(20, 184, 166, 0.18);
-  --constellation-dot: rgba(20, 184, 166, 0.45);
-  --constellation-glow: rgba(20, 184, 166, 0.08);
+  --constellation-line: rgba(15, 118, 110, 0.22);
+  --constellation-dot: rgba(15, 118, 110, 0.45);
+  --constellation-glow: rgba(15, 118, 110, 0.08);
 }
 
 body {
